python/image_converter/requirements.txt
opencv-python==4.5.5.64
Pillow>=9.0.1
